+++ core/src/test/resources/fixtures/manifest-groovy-java/app/src/test/java/com/example/oldname/ExampleUnitTest.java
@@ -1,0 +1,12 @@
+package com.example.oldname;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertTrue;
+
+public class ExampleUnitTest {
+    @Test
+    public void test() {
+        assertTrue(true);
+    }
+}